Add files via upload
Define a class named 'triangle' to represent a triangle using the lengths of the three sides. Write a constructor to initialize objects of this class, given the lengths of the sides. Also write member functions to check 
(a) if a triangle is isosceles 
(b) if a triangle is equilateral
Write a main function to test your functions