Add files via upload
Define a class " point " with two data members " xordinate " and " yordinate " to represent all points in the two dimensional plane by storing their x co-ordinate and y co-ordinate values. Write member functions 
dist( ): to return the distance of the point from the origin 
slope( ): to return the slope of the line obtained by joining this point with the origin. 
Write constructors with zero, one, and two arguments to initialize objects. Also Write a friend function to compute the distance between two points.