Add files via upload
Define a abstract base class figure and add to it pure virtual functions
display() : to display a figure
get() : to input parameters of the figure
area() : to compute area of a figure
perimeter() : to compute the perimeter of a figure
Drive three classes circle, rectangle, triangle from it. A  circle is to be represented by its radius, rectangle by its length and breadth and triangle by its lengths of their sides. Write a main function and write necessary statements to achieve run time polymorphism.